Hunt and Kill phase 1
I've successfully implemented the first few steps in the hunt and kill algorithm. All that needs to be done now is to have it repeat the process until all cells have been visited. The code may not be too pretty right now, however one of the next things on my todo list is to clean it up.