Maze Generation Fixes. I was a fool :^)
For some reason, I had been under the impression that arrays in BlitzBasic could not be multidimensional. This misunderstanding, due in part to my misreading of the command reference, had caused a great deal of frustration. Thankfully, I've discovered that I was wrong and have made several changes to re-implement the maze generator with this new knowledge in mind.