Maze Generation Working
The functions for generating mazes have been completed and the maze is being created as expected. The main takeaways from this commit are the inclusion of a mode parameter to 'checkCell', the completion of the 'hunt' function, and the inclusion of the 'loadMessage' function. The additional parameter in 'checkCell' was added primarily so I could avoid creating unnecessary 'if' statements in 'hunt' and so I could re-use some code. As for 'loadMessage', it was added in response to improper maze generation due to internal clock speed (At least, I suspect). Basically, I added 'loadMessage' to slow down program execution enough to generate the maze right.