Collisions, Exits, and Textures
I've setup the mesh for the exit and assigned it to a global variable. I've also added a texture the maze walls and turned on their collision. Aside from that I've added a few minor tweaks to the controls and camera. One thing to note is the removal of the 'cell' type, this was due to it not being needed as I could just make the maze walls part of the ground.