Menu and Minor Changes
I've added a menu, though not as pretty as before, and the ability to choose the size of the maze through the difficulty. Along with this, I have changed the order of things a bit and removed some commented out code. Maybe in the future I'll make the menu a bit more pleasing to look at, though it's fine for now.